Create Meryl July 2021
This GPAO calculation takes the existing term GPA and subtracts out the (weighted by units) class grade
Rounding errors in ucd_term_gpa are a problem. I use the ceil() command to round up to the nearest integer. This isn't perfect, but it does prevent negative GPAO! It will result in some erroneously high GPAOs- but not by much.
The advantage of this approach as opposed to trying to calculate from scratch is that we can make sure not to include classes that shouldn't count for GPA